# docs/ApiSpec-openapi.yaml
...
openapi: 3.0.1
info:
  title: OcrProject
  version: 1.0.0 
paths:
...
  /documents/search:
    get:
      operationId: searchDocuments
      summary: Search documents
      description: Users can search via arbitrary keywords to return a list of relevant documents
      parameters:
      - name: query
        in: query
        description: The keyword(s) to search for
        required: true
        schema:
          type: string
      - name: pageSize
        in: query
        description: The maximum number of items to return
        required: false
        schema:
          default: 10
          maximum: 50
          type: integer
      - name: page
        in: query
        description: The page number to load
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'
...
  /documents/{id}/image:
    get:
      operationId: getDocumentImage
      summary: Get document image
      description: Get original scanned pages for the document
      parameters:
      - name: id
        in: path
        description: 'The id of the document'
        required: true
        schema:
          type: integer
      - name: page
        in: query
        description: 'The page to load'
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: envelope
        in: query
        description: 'This flag parameter indicates that the image of the envelope should be returned, instead of a page. This field is mutually exclusive with the page parameter. If both are specified,
                      then the envelope will be returned and the page is ignored..'
        required: false
      responses:
        200:
          description: OK
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/tiff:
              schema:
                type: string
                format: binary
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    DisputeQueue:
      type: string
      enum:
        - FRAUD
        - VALIDATION
        - GENERAL
        - INQUIRIES
        - UNASSIGNED
...
    Error:
      description: 'This object is returned when an error (non-2xx response codde)
        occurs. '
      required:
      - message
      - status
      properties:
        message:
          description: A brief message describing the error
          type: string
        status:
          description: A copy of the HTTP status code returned by the request.
          type: integer
      example: |-
        {
          "message": "Some message",
          "status": 123
        }
    Address:
      description: 'A street address'
      required:
        - firstLine
        - city
        - state
        - country
      properties:
        firstLine:
          type: string
        secondLine:
          type: string
        city:
          type: string
        state:
          description: "State/province/region"
          type: string
        zipCode:
          type: string
        country:
          type: string
      example: |-
        {
          firstLine: "1600 Pennsylvania Ave.",
          state: "Washington D.C.",
          zipCode: "20500"
        }
    DocumentResponse:
      description: Metadata that has been extracted from an uploaded document'
      properties:
        id:
          description: The name of the document
          type: integer
        accountNumber:
          description: The account number related to this document
          type: string
        address:
          $ref: '#/components/schemas/Address'
        ssn:
          type: string
          pattern: '^[0-9]{3}-[0-9]{2}-[0-9]{4}$'
        dateOfBirth:
          description: 'A UNIX timestamp in seconds representing a date'
          type: integer
        letterDate:
          description: 'A UNIX timestamp in seconds representing a date'
          type: integer
        postmarkDate:
          description: 'A UNIX timestamp in seconds representing a date'
          type: integer

        numSimilarDocuments:
          description: "The number of documents that are substantially similar to this one. A high number here indicates that 
                        the document is likely based off of a template."
          type: integer
          minimum: 0
        numPages:
          description: 'The number of pages in the document'
          type: integer
          minimum: 0
        hasEnvelope:
          type: boolean
        disputeQueue:
          description: "The dispute queue to assign the document to."
          type: string
          oneOf:
            - $ref: '#/components/schemas/DisputeQueue'
...